# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClassDefinition {
  slug: string;
  label: string;
  description: string;
  targetCount: number;
}

export interface DatasetStatus {
  totalImages: number;
  trainImages: number;
  validationImages: number;
  testImages: number;
  isReady: boolean;
  readinessLabel: string;
  split: string;
}

export interface TrainingMetrics {
  /** @nullable */
  accuracy: number | null;
  /** @nullable */
  precision: number | null;
  /** @nullable */
  recall: number | null;
  /** @nullable */
  epochs: number | null;
  /** @nullable */
  lastRun: string | null;
}

export interface ClassifierOverview {
  projectName: string;
  model: string;
  modelFamily: string;
  modelRationale: string;
  inputSize: number;
  dataset: DatasetStatus;
  classes: ClassDefinition[];
  metrics: TrainingMetrics;
  inferenceReady: boolean;
  inferenceMessage: string;
}

export interface PredictionInput {
  /**
     * Base64-encoded image data URL.
     * @minLength 32
     */
  imageData: string;
  fileName?: string;
}

export type PredictionAlternativesItem = {
  label: string;
  confidence: number;
};

export interface Prediction {
  label: string;
  confidence: number;
  alternatives: PredictionAlternativesItem[];
  model: string;
}

export interface ErrorResponse {
  error: string;
  code: string;
}