# lib/api-zod/src/generated/types/classDefinition.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClassDefinition {
  slug: string;
  label: string;
  description: string;
  targetCount: number;
}