# lib/api-zod/src/generated/types/classifierOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClassDefinition } from './classDefinition';
import type { DatasetStatus } from './datasetStatus';
import type { TrainingMetrics } from './trainingMetrics';

export interface ClassifierOverview {
  projectName: string;
  model: string;
  modelFamily: string;
  modelRationale: string;
  inputSize: number;
  dataset: DatasetStatus;
  classes: ClassDefinition[];
  metrics: TrainingMetrics;
  inferenceReady: boolean;
  inferenceMessage: string;
}