# lib/api-zod/src/generated/types/datasetStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DatasetStatus {
  totalImages: number;
  trainImages: number;
  validationImages: number;
  testImages: number;
  isReady: boolean;
  readinessLabel: string;
  split: string;
}