# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionAlternativesItem } from './predictionAlternativesItem';

export interface Prediction {
  label: string;
  confidence: number;
  alternatives: PredictionAlternativesItem[];
  model: string;
}