# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  /**
     * Base64-encoded image data URL.
     * @minLength 32
     */
  imageData: string;
  fileName?: string;
}