# lib/api-zod/src/generated/types/trainingMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrainingMetrics {
  /** @nullable */
  accuracy: number | null;
  /** @nullable */
  precision: number | null;
  /** @nullable */
  recall: number | null;
  /** @nullable */
  epochs: number | null;
  /** @nullable */
  lastRun: string | null;
}